build: define upstream-repos in modules-file
Switch from defining the upstream-repos and relating commit in config.mk
and feeds.conf to define all in the modules-file.
The modules-files is the real file used by the build-system to setup the
working checkout. It originates from the Gluon-buildsystemand some
documentation is avail at https://gluon.readthedocs.io/en/latest/dev/build.html

This also relates to
* PR https://github.com/freifunk-berlin/firmware/pull/781
* Issue https://github.com/freifunk-berlin/firmware/issues/780
* Issue https://github.com/freifunk-berlin/firmware/issues/772