build: use separate Makefile for CI and drop IS_BUILDBOT
Add a separte Makefile that will be used for building in a CI-environment.
This new Makefile.autobuild is just some glue to provide simple make-
targets which will then call the corresponding make-target(s) of the "real"
Makefile. The new file will also directly deal with the special cases we
control with the "IS_BUILDBOT" and the "SET_BUILDBOT" envs / flags. So we
can get rid of them again and two more easy to read Makefiles without
additional "#if" macros.