profiles: drop Ubnt LS-SR71
This is an embedded board for a miniPCI socket, which was added when switching
to LEDE buildbasis (a05e97b6ac1333b0c710d6de65f762ee57642a07). As we focus on
regular off-the-shelf routers, we stop to build this image.
On the statistics there is also no indication, that this board is used anyway.