patches/openwrt: restore status of services after sysupgrade
after a sysupgrade disabled services are reenabled which is likely not expected
behavior.
So on a sysupgrade also keep the list of installed packages and include the
status of the services. after sysupgrading the new firmware, run a uci-default
script to restore the status.

This change was send upstream, but rejected as not being generic enough [1].
Let's keep it for our purpose and rework later.

1 - http://lists.openwrt.org/pipermail/openwrt-devel/2021-January/033270.html