configs: revert to upstreams idea of SQUASHFS_BLOCK_SIZE (beside for tiny devices)
back in e6708dead14c6151a588e724dc489b486468e0de the SQUASHFS_BLOCK_SIZE was set to 512 to provide
better compression, esp. for 4MB flash devices. When OpenWrt setup the tiny subtarget
for such devices, where we keep this special handling. For all other subtargets we can
revert back to the upstream-default of this setting (256 kb in April 2021).