Revert "don't build sqaushfs-images during build-time"
The idea is working and saves build-time and build-space, but the generated imagebuilder
will also have the "no squashfs" embedded. This is done by copying the build-time .config
file.
Reenable buildung the squashfs-images can be dont with a small pach inside teh imagebuilder
or in teh assembel_firmwares.sh script. But we are missing then the SQUASHFS_BLOCK_SIZE and
there seems no easy way to lookup the OpenWrt default, so we have two options:
* hardcoding a own value
* building some "magic" to extract the default from Config.in

Both options don't seem really great, so drop the whole idea of "avoiding the squashfs
during buildtime".

* Revert "Squashfs: try to lookup SQUASHFS_BLOCK_SIZE-during-buildtime"
* Revert "Squashfs: reenable TARGET_ROOTFS_SQUASHFS on imagecreation"
* Revert "configs: stop building squashfs images during build-phase"