patches/openwrt: drop upstreamed "tools-mkimake-fix-handling-of-long-pathnames"
When OpenWrt switched to U-Boot 2021.01 in e6ccb40ba5bedbc4a27ddea4a858f17c3496a43f this
patch became obsolete, as it was integrated in this upstream-version.
However quilt was able to apply the patch, by adding it twice, so it did not break
the build. But it's safe now to be dropped.